Applied Limits to Template Matching
Not perfect, but removed the normalize function from template matching,
and this allowed the results to be different enough to filter.  Now
values can be seen when viewing the matching in action, so a range of
acceptable values can be applied after some testing. This will allow for
showing objects with high similarities to the original target