Improved drone TCP server and client
TCP server now can receive frames from both clients. Only the "active"
drone will be sending frames to the server, while the "dormant" drone
will wait to send it's frames to the server. These frames for only the
"active" drone will be displayed. The server for sending movement
commands to the client drones remains the same.

TCP client now sends frames to the server. The client waits to
synchronize with the server, making sure it's frames are not sent if it
is not considered the "active" drone, saving valuable CPU resources.
Once the client knows it is "active", it synchronizes again with the
drone for each frame being sent, confirming that the server successfully
received the frame size. After this, the client sends over the frame to
the server. Last, there is a final synchronization with the server for
each frame sent to validate that the server has received the entire
frame successfully.