Sass Class - Main for final project compilation & run
Created a Sass class, which will be used to run the project as a whole.
The runSystem function is the main function. This class contains the
two server function/threads for receiving frames from the drones and
sending commands to them. It also contains a thread on which the drone
objects will analyze their respective image frames and decide whether
there is a command that needs to be sent. The send command thread will
constantly be checking the public Communication Box to see if there are
commands waiting to be sent.