Communication Box
Struct meant to hold all of the communication-related data across all of
the threads being maintained by the Sass class.  The latest video frames
are added to this struct. The commands that are to be sent from each
drone object to the sending thread are added to this struct, and there's
a boolean wait variable for each that will keep the command-sending
system in order