Updated functions sass.cpp
Updated and implemented the receive frames function as well as the send commands function to make sure they are complete. More revision will happen later on. 

Side Note: Still having memory problems in the receive frame function. which causes a core dump, because the read() is returning an EFAULT (bad address)