Drone template matching functionality
Added the ability to add images of a scanned human in a scaled down form
to the drone's memory, so that it can recognize the human in its camera
feed. There is a boolean for displaying template matching, or turning it
off.  Template matching is not used for decision making, just visuals.